+++ docs/news.rst
@@ -8,6 +8,7 @@ Lab
 ^^^
 * Automatically group multiple runs into one Slurm task when the number
   of runs exceeds the maximum number of Slurm tasks (Jendrik Seipp).
+* Add ``time_limit_per_task`` parameter to ``SlurmEnvironment``.
 
 Downward Lab
 ^^^^^^^^^^^^